(GH-2302) Moves assembly resolver to outer scope.
- Rename ScriptAssemblyResolver to CakeAssemblyResolver
- Omit .resources in assembly resolver
- Fixes #2302
